+++ CHANGELOG.md
@@ -6,7 +6,10 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## [UNRELEASED]
 ### Fixed
-- NaN
+- Extended UTF-7 support added (RFC2060) #383
+- IMAP Quota root command fixed
+- Prevent line-breaks in folder path caused by special chars
+- Partial fix for #362 (allow overview response to be empty)
 
 ### Added
 - `Protocol::sizes()` support added (fetch the message byte size via RFC822.SIZE). Accessible through `Message::getSize()` #379 (thanks @didi1357)